+++ tests/mpd/protocol/__init__.py
@@ -6,9 +6,8 @@ import unittest
 import pykka
 
 from mopidy import core
+from mopidy.backend import dummy
 from mopidy.mpd import session
-
-from tests import dummy_backend
 
 
 class MockConnection(mock.Mock):
@@ -32,7 +31,7 @@ class BaseTestCase(unittest.TestCase):
         }
 
     def setUp(self):
-        self.backend = dummy_backend.create_dummy_backend_proxy()
+        self.backend = dummy.create_dummy_backend_proxy()
         self.core = core.Core.start(backends=[self.backend]).proxy()
 
         self.connection = MockConnection()