+++ tests/utils/test_jsonrpc.py
@@ -7,9 +7,8 @@ import unittest
 import pykka
 
 from mopidy import core, models
+from mopidy.backend import dummy
 from mopidy.utils import jsonrpc
-
-from tests import dummy_backend
 
 
 class Calculator(object):
@@ -41,7 +40,7 @@ class Calculator(object):
 
 class JsonRpcTestBase(unittest.TestCase):
     def setUp(self):
-        self.backend = dummy_backend.create_dummy_backend_proxy()
+        self.backend = dummy.create_dummy_backend_proxy()
         self.core = core.Core.start(backends=[self.backend]).proxy()
 
         self.jrw = jsonrpc.JsonRpcWrapper(